refactor(sdk): extract stream-turn + context-object orchestrator constructor
three mechanical cuts on runtime/query/iteration/index.ts (1076
lines before the busy-wait teardown), zero behavior change:

1. streamProviderTurn (~330 lines) plus its private helpers
   (synthesizeMessageStopReason, StreamingTurnResult) move to
   iteration/stream-turn.ts, a sibling of phases/. the function
   body is byte-identical to the original — verified by diffing
   the extracted block against the pre-refactor file.

2. IterationOrchestrator's 12 positional constructor args and the
   IterationConfig bag collapse into the already-existing
   IterationContext shape: constructor(ctx: IterationContext).
   the sole call site (runtime/query/index.ts) now builds the
   context object; the launchedTasks ?? new Map() default moves
   with it. IterationConfig had no other importers and is not
   part of the package export surface, so it is deleted.

3. AUTO_CONTINUATION_USER_MESSAGE moves from
   runtime/query/continuation.ts to constants/continuation.ts
   (re-exported via constants/index.ts). manager/run/persistence.ts
   consumed it through a manager -> runtime import, the wrong
   dependency direction for a leaf string constant; both importers
   now read it from constants/ and the old module is deleted,
   breaking the manager<->runtime directory cycle.

sdk test suite ran green after each step.
