+++ packages/sdk/src/constants/continuation.ts
@@ -11,6 +11,11 @@
  * that match this constant verbatim, so the run's `result` field
  * concatenates the full multi-turn assistant output instead of only
  * surfacing the trailing continuation chunk.
+ *
+ * Lives in `constants/` (not `runtime/query/`) because both the
+ * runtime iteration loop and `manager/run/persistence.ts` consume
+ * it — a manager→runtime import for a string constant would create
+ * a manager↔runtime directory cycle.
  */
 export const AUTO_CONTINUATION_USER_MESSAGE =
 	'Continue exactly where you left off. Do not repeat content you already wrote — pick up at the next token.'
