+++ packages/sdk/src/constants/index.ts
@@ -7,6 +7,7 @@ export * from './rag/index.js'
 export * from './tools/index.js'
 export * from './telemetry/index.js'
 export * from './agent/index.js'
+export * from './continuation.js'
 export * from './limits.js'
 export * from './bus/index.js'
 export * from './verification/index.js'
